+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Utility/CryptogramUtility.cs
@@ -8,13 +8,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region Offset
         public static void OffsetEncryptBundleFiles(string dir, int dummySize)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.Offset.WriteFile.EncryptFile(fPath, dummySize);
             }
@@ -22,13 +22,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void OffsetDecryptBundleFiles(string dir, int dummySize)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.Offset.WriteFile.DecryptFile(fPath, dummySize);
             }
@@ -38,13 +38,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region Xor
         public static void XorEncryptBundleFiles(string dir, byte key)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.XOR.WriteFile.EncryptFile(fPath, key);
             }
@@ -52,13 +52,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void XorDecryptBundleFiles(string dir, byte key)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.XOR.WriteFile.DecryptFile(fPath, key);
             }
@@ -68,13 +68,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region HT2Xor
         public static void HT2XorEncryptBundleFiles(string dir, byte hKey, byte tKey, byte jKey)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.HT2XOR.WriteFile.EncryptFile(fPath, hKey, tKey, jKey);
             }
@@ -82,13 +82,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void HT2XorDecryptBundleFiles(string dir, byte hKey, byte tKey, byte jKey)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.HT2XOR.WriteFile.DecryptFile(fPath, hKey, tKey, jKey);
             }
@@ -98,13 +98,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region HT2XorPlus
         public static void HT2XorPlusEncryptBundleFiles(string dir, byte hKey, byte tKey, byte j1Key, byte j2key)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.HT2XORPlus.WriteFile.EncryptFile(fPath, hKey, tKey, j1Key, j2key);
             }
@@ -112,13 +112,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void HT2XorPlusDecryptBundleFiles(string dir, byte hKey, byte tKey, byte j1Key, byte j2key)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.HT2XORPlus.WriteFile.DecryptFile(fPath, hKey, tKey, j1Key, j2key);
             }
@@ -128,13 +128,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region AES
         public static void AesEncryptBundleFiles(string dir, string key, string iv)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.AES.WriteFile.EncryptFile(fPath, key, iv);
             }
@@ -142,13 +142,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void AesDecryptBundleFiles(string dir, string key, string iv)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.AES.WriteFile.DecryptFile(fPath, key, iv);
             }
@@ -158,13 +158,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region ChaCha20
         public static void ChaCha20EncryptBundleFiles(string dir, string key, string nonce, uint counter)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.ChaCha20.WriteFile.EncryptFile(fPath, key, nonce, counter);
             }
@@ -172,13 +172,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void ChaCha20DecryptBundleFiles(string dir, string key, string nonce, uint counter)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.ChaCha20.WriteFile.DecryptFile(fPath, key, nonce, counter);
             }
@@ -188,13 +188,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region XXTEA
         public static void XXTEAEncryptBundleFiles(string dir, string key)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.XXTEA.WriteFile.EncryptFile(fPath, key);
             }
@@ -202,13 +202,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void XXTEADecryptBundleFiles(string dir, string key)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.XXTEA.WriteFile.DecryptFile(fPath, key);
             }
@@ -218,13 +218,13 @@ namespace OxGFrame.AssetLoader.Utility
         #region OffsetXOR
         public static void OffsetXorEncryptBundleFiles(string dir, byte key, int dummySize)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行加密
+            // 對所有文件進行加密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的加密
+                // 執行各文件的加密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.OffsetXOR.WriteFile.EncryptFile(fPath, key, dummySize);
             }
@@ -232,13 +232,13 @@ namespace OxGFrame.AssetLoader.Utility
 
         public static void OffsetXorDecryptBundleFiles(string dir, byte key, int dummySize)
         {
-            // 取得目錄下所有檔案
+            // 取得目錄下所有文件
             FileInfo[] files = BundleUtility.GetFilesRecursively(dir);
 
-            // 對所有檔案進行解密
+            // 對所有文件進行解密
             for (int i = 0; i < files.Length; i++)
             {
-                // 執行各檔案的解密
+                // 執行各文件的解密
                 string fPath = Path.Combine(files[i].Directory.ToString(), files[i].Name);
                 FileCryptogram.OffsetXOR.WriteFile.DecryptFile(fPath, key, dummySize);
             }