+++ Assets/OxGFrame/AssetLoader/Tests/Scripts/Editor/Bundle/ChaCha20Tests.cs
@@ -3,7 +3,6 @@ using OxGFrame.AssetLoader.Utility;
 using System;
 using System.Diagnostics;
 using System.IO;
-using System.Threading;
 using static OxGFrame.AssetLoader.Bundle.FileCryptogram;
 
 namespace OxGFrame.AssetLoader.Editor.Tests
@@ -14,21 +13,19 @@ namespace OxGFrame.AssetLoader.Editor.Tests
         internal readonly string nonce = "nrKpHnV7F1";
         internal readonly uint counter = 2;
 
-        internal readonly ulong dataSize = 1024 * 1024;
-
         [Test]
         public void EncryptDecryptBytesFromData()
         {
             Stopwatch stopwatch = new Stopwatch();
 
-            byte[] testBytes = new byte[dataSize];
+            byte[] testBytes = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testBytes);
             byte[] originalBytes = (byte[])testBytes.Clone();
 
             stopwatch.Start();
             bool encryptResult = ChaCha20.EncryptBytes(ref testBytes, key, nonce, counter);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] ChaCha20.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] ChaCha20.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(encryptResult, "In-place encryption failed");
 
             stopwatch.Reset();
@@ -36,7 +33,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             bool decryptResult = ChaCha20.DecryptBytes(ref testBytes, key, nonce, counter);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] ChaCha20.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] ChaCha20.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(decryptResult, "In-place decryption failed");
 
             Assert.AreEqual(originalBytes, testBytes, "Decrypted content does not match the original content");
@@ -48,14 +45,14 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
             stopwatch.Start();
             bool encryptResult = ChaCha20.WriteFile.EncryptFile(tempFile, key, nonce, counter);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] ChaCha20.WriteFile.EncryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] ChaCha20.WriteFile.EncryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(encryptResult, "File encryption failed");
 
             stopwatch.Reset();
@@ -63,7 +60,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             bool decryptResult = ChaCha20.WriteFile.DecryptFile(tempFile, key, nonce, counter);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] ChaCha20.WriteFile.DecryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] ChaCha20.WriteFile.DecryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(decryptResult, "File decryption failed");
 
             byte[] decryptedData = File.ReadAllBytes(tempFile);
@@ -78,14 +75,14 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
             stopwatch.Start();
             byte[] encryptedBytes = ChaCha20.EncryptBytes(tempFile, key, nonce, counter);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] ChaCha20.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] ChaCha20.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsNotNull(encryptedBytes, "Encrypted bytes returned null");
             Assert.IsNotEmpty(encryptedBytes, "Encrypted bytes are empty");
 
@@ -97,7 +94,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             byte[] decryptedBytes = ChaCha20.DecryptBytes(encryptedFile, key, nonce, counter);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] ChaCha20.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] ChaCha20.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsNotNull(decryptedBytes, "Decrypted bytes returned null");
             Assert.AreEqual(testData, decryptedBytes, "Decrypted content does not match the original content");
 
@@ -111,7 +108,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
@@ -131,7 +128,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
                     Assert.AreEqual(testData, decryptedData, "Stream decrypted content does not match");
                 }
             }
-            UnityEngine.Debug.Log($"[DecryptStream] ChaCha20.DecryptStream execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[DecryptStream] ChaCha20.DecryptStream execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
 
             File.Delete(tempFile);
         }