+++ Assets/OxGFrame/AssetLoader/Tests/Scripts/Editor/Bundle/CryptogramConfig.cs
@@ -1,0 +1,7 @@
+namespace OxGFrame.AssetLoader.Editor.Tests
+{
+    internal class CryptogramConfig
+    {
+        public const ulong DATA_SIZE = 1024 * 1024;
+    }
+}