+++ Assets/OxGFrame/AssetLoader/Tests/Scripts/Editor/Bundle/HT2XORPlusTests.cs
@@ -3,7 +3,6 @@ using OxGFrame.AssetLoader.Utility;
 using System;
 using System.Diagnostics;
 using System.IO;
-using System.Threading;
 using static OxGFrame.AssetLoader.Bundle.FileCryptogram;
 
 namespace OxGFrame.AssetLoader.Editor.Tests
@@ -15,21 +14,19 @@ namespace OxGFrame.AssetLoader.Editor.Tests
         internal readonly byte j1Key = 3;
         internal readonly byte j2Key = 4;
 
-        internal readonly ulong dataSize = 1024 * 1024;
-
         [Test]
         public void EncryptDecryptBytesFromData()
         {
             Stopwatch stopwatch = new Stopwatch();
 
-            byte[] testBytes = new byte[dataSize];
+            byte[] testBytes = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testBytes);
             byte[] originalBytes = (byte[])testBytes.Clone();
 
             stopwatch.Start();
             bool encryptResult = HT2XORPlus.EncryptBytes(testBytes, hKey, tKey, j1Key, j2Key);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] HT2XORPlus.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] HT2XORPlus.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(encryptResult, "In-place encryption failed");
 
             stopwatch.Reset();
@@ -37,7 +34,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             bool decryptResult = HT2XORPlus.DecryptBytes(testBytes, hKey, tKey, j1Key, j2Key);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] HT2XORPlus.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] HT2XORPlus.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(decryptResult, "In-place decryption failed");
 
             Assert.AreEqual(originalBytes, testBytes, "Decrypted content does not match the original content");
@@ -49,14 +46,14 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
             stopwatch.Start();
             bool encryptResult = HT2XORPlus.WriteFile.EncryptFile(tempFile, hKey, tKey, j1Key, j2Key);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] HT2XORPlus.WriteFile.EncryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] HT2XORPlus.WriteFile.EncryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(encryptResult, "File encryption failed");
 
             stopwatch.Reset();
@@ -64,7 +61,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             bool decryptResult = HT2XORPlus.WriteFile.DecryptFile(tempFile, hKey, tKey, j1Key, j2Key);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] HT2XORPlus.WriteFile.DecryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] HT2XORPlus.WriteFile.DecryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(decryptResult, "File decryption failed");
 
             byte[] decryptedData = File.ReadAllBytes(tempFile);
@@ -79,14 +76,14 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
             stopwatch.Start();
             byte[] encryptedBytes = HT2XORPlus.EncryptBytes(tempFile, hKey, tKey, j1Key, j2Key);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] HT2XORPlus.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] HT2XORPlus.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsNotNull(encryptedBytes, "Encrypted bytes returned null");
             Assert.IsNotEmpty(encryptedBytes, "Encrypted bytes are empty");
 
@@ -98,7 +95,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             byte[] decryptedBytes = HT2XORPlus.DecryptBytes(encryptedFile, hKey, tKey, j1Key, j2Key);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] HT2XORPlus.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] HT2XORPlus.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsNotNull(decryptedBytes, "Decrypted bytes returned null");
             Assert.AreEqual(testData, decryptedBytes, "Decrypted content does not match the original content");
 
@@ -112,7 +109,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
@@ -132,7 +129,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
                     Assert.AreEqual(testData, decryptedData, "Stream decrypted content does not match");
                 }
             }
-            UnityEngine.Debug.Log($"[DecryptStream] HT2XORPlus.DecryptStream execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[DecryptStream] HT2XORPlus.DecryptStream execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
 
             File.Delete(tempFile);
         }