+++ Assets/OxGFrame/AssetLoader/Tests/Scripts/Editor/Bundle/OffsetTests.cs
@@ -3,7 +3,6 @@ using OxGFrame.AssetLoader.Utility;
 using System;
 using System.Diagnostics;
 using System.IO;
-using System.Threading;
 using static OxGFrame.AssetLoader.Bundle.FileCryptogram;
 
 namespace OxGFrame.AssetLoader.Editor.Tests
@@ -12,21 +11,19 @@ namespace OxGFrame.AssetLoader.Editor.Tests
     {
         internal readonly int dummySize = 128;
 
-        internal readonly ulong dataSize = 1024 * 1024;
-
         [Test]
         public void EncryptDecryptBytesFromData()
         {
             Stopwatch stopwatch = new Stopwatch();
 
-            byte[] testBytes = new byte[dataSize];
+            byte[] testBytes = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testBytes);
             byte[] originalBytes = (byte[])testBytes.Clone();
 
             stopwatch.Start();
             bool encryptResult = Offset.EncryptBytes(ref testBytes, dummySize);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] Offset.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] Offset.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(encryptResult, "In-place encryption failed");
 
             stopwatch.Reset();
@@ -34,7 +31,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             bool decryptResult = Offset.DecryptBytes(ref testBytes, dummySize);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] Offset.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromData] Offset.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(decryptResult, "In-place decryption failed");
 
             Assert.AreEqual(originalBytes, testBytes, "Decrypted content does not match the original content");
@@ -46,14 +43,14 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
             stopwatch.Start();
             bool encryptResult = Offset.WriteFile.EncryptFile(tempFile, dummySize);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] Offset.WriteFile.EncryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] Offset.WriteFile.EncryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(encryptResult, "File encryption failed");
 
             stopwatch.Reset();
@@ -61,7 +58,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             bool decryptResult = Offset.WriteFile.DecryptFile(tempFile, dummySize);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] Offset.WriteFile.DecryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptWriteFile] Offset.WriteFile.DecryptFile execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsTrue(decryptResult, "File decryption failed");
 
             byte[] decryptedData = File.ReadAllBytes(tempFile);
@@ -76,14 +73,14 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
             stopwatch.Start();
             byte[] encryptedBytes = Offset.EncryptBytes(tempFile, dummySize);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] Offset.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] Offset.EncryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsNotNull(encryptedBytes, "Encrypted bytes returned null");
             Assert.IsNotEmpty(encryptedBytes, "Encrypted bytes are empty");
 
@@ -95,7 +92,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             stopwatch.Start();
             byte[] decryptedBytes = Offset.DecryptBytes(encryptedFile, dummySize);
             stopwatch.Stop();
-            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] Offset.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[EncryptDecryptBytesFromFile] Offset.DecryptBytes execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
             Assert.IsNotNull(decryptedBytes, "Decrypted bytes returned null");
             Assert.AreEqual(testData, decryptedBytes, "Decrypted content does not match the original content");
 
@@ -109,7 +106,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
             Stopwatch stopwatch = new Stopwatch();
 
             string tempFile = Path.GetTempFileName();
-            byte[] testData = new byte[dataSize];
+            byte[] testData = new byte[CryptogramConfig.DATA_SIZE];
             new Random().NextBytes(testData);
             File.WriteAllBytes(tempFile, testData);
 
@@ -128,7 +125,7 @@ namespace OxGFrame.AssetLoader.Editor.Tests
                     Assert.AreEqual(testData, decryptedData, "Stream decrypted content does not match");
                 }
             }
-            UnityEngine.Debug.Log($"[DecryptStream] Offset.DecryptStream execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, DataSize: {BundleUtility.GetBytesToString(dataSize)}");
+            UnityEngine.Debug.Log($"[DecryptStream] Offset.DecryptStream execution time: {stopwatch.Elapsed.TotalMilliseconds} ms, CryptogramConfig.DATA_SIZE: {BundleUtility.GetBytesToString(CryptogramConfig.DATA_SIZE)}");
 
             File.Delete(tempFile);
         }